refactor: create keylogger usecase
