--- src/main/java/infra/composers/KeyloggerControllerFactory.java
@@ -1,14 +1,0 @@
-package infra.composers;
-
-import java.util.function.Consumer;
-import adapters.keyloggerController.KeyloggerController;
-import infra.keylistenner.Keylistenner;
-
-public class KeyloggerControllerFactory {
-    static public KeyloggerController create(Consumer<String> onKeyPressed) {
-        return new KeyloggerController(
-            new Keylistenner(),
-            onKeyPressed
-        );
-    }
-}
