+++ src/main/java/infra/composers/UsecaseFactory.java
@@ -1,9 +1,12 @@
 package infra.composers;
 
+import java.util.function.Consumer;
+
 import infra.keylistenner.Keylistenner;
 import infra.repositories.shortcutsRepository.ShortcutsRepository;
 import infra.robot.Robot;
 import usecases.keyEventsScanner.KeyEventsScanner;
+import usecases.keyLoggerRunner.KeyLoggerRunner;
 import usecases.shortcutUpdater.ShortcutUpdater;
 import usecases.shortcutsRunner.ShortcutsRunner;
 
@@ -22,5 +25,12 @@ public class UsecaseFactory {
 
     static public ShortcutUpdater createShortcutUpdater() {
         return new ShortcutUpdater(new ShortcutsRepository());
-    } 
+    }
+
+    static public KeyLoggerRunner createKeyLoggerRunner(Consumer<String> onKeyPressed) {
+        return new KeyLoggerRunner(
+            new Keylistenner(),
+            onKeyPressed
+        );
+    }
 }
